+++ CHANGELOG.md
@@ -5,6 +5,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.13.2] Q3 2023
+- bumped to azure-iot-sdk 0.11.0 which prevents potential deadlocks 
+- introduced signal handler to handle termination and shutdown twin properly
+
 ## [0.13.1] Q3 2023
 - fixed rust-toolchain.toml